ConferenceV2/requirements.txt
fastapi
uvicorn
azure-messaging-webpubsubservice
python-dotenv
pydantic
aiohttp
azure-cosmos
python-dotenv
